renamed package from vexpo.virtualexpo to nammaexpo
