+++ src/main/java/com/nammaexpo/VirtualExpoApplication.java
@@ -1,4 +1,4 @@
-package com.vexpo.virtualexpo;
+package com.nammaexpo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
