+++ src/main/java/com/nammaexpo/swagger/SwaggerConfig.java
@@ -1,0 +1,42 @@
+package org.sid.tool.swagger;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+import static springfox.documentation.builders.PathSelectors.regex;
+
+/**
+ * Class is to enable swagger documentation
+ *
+ * @author siddesh
+ * @since 09/jan/2020
+ */
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {
+
+    @Bean
+    public Docket productApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("org.sid.tool"))
+                .paths(regex("/.*"))
+                .build()
+                .apiInfo(metaData());
+    }
+
+    private ApiInfo metaData() {
+        return new ApiInfo(
+                "Spring Boot REST API", "Spring Boot REST API for Prject Management Tool",
+                "1.0",
+                "Terms of service",
+                "Siddesh",
+                "Apache License Version 2.0",
+                "https://www.apache.org/licenses/LICENSE-2.0");
+    }
+}
