+++ src/test/java/com/nammaexpo/VirtualExpoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.vexpo.virtualexpo;
+package com.nammaexpo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
